Applied patch [ 587454 ] CW-related wx/wxchar.h fix
This fixes the problem in include/wx/wxchar.h with
including <wchar.h> for CW reported by Andreas
Simanowski.

The current version does not work with CW Win32 builds.

The fix suggested by Andreas is incorrect, because it
will break the wxMAC CW builds.

The purpose of conditionals starting at line 77 is to
correctly handle the various compilers. The fundamental
problem with the existing code is that it begins by
conditioning on __WIN32__. This is wrong, and makes
it hard to handle compilers that work on more than one
OS (such as CW).

Instead the conditionals should be done purely on
compilers. Each compiler can then set things
appropriately. If a particular compiler works on more
than one OS but has different needs on each OS, then
this can be handled within the code for that compiler.

The attached patch removes the top level conditional on
__WIN32__ and restructures the remainder as a series
of conditionals on individual compilers.

Igor Mikolic-Torreira
igormt@alumni.caltech.edu