Applied [ 587500 ] Update CW project file for wxMSW
This patch updates the CW project file for building
wxMSW.

It patches the file src/wxWindowsW7.xml on the
development branch.

It does the following:

(1) adds the file src/common/msgout.cpp to the build.

(2) moves popupwin.cpp ahead of sub-libraries in the link
order for linking the library (this is just paranoia on my
part ;)

Igor Mikolic-Torreira
igormt@alumni.caltech.edu