Applied patch [ 584057 ] Fixes calculation bug in wxThread::Sleep
There's a problem in src/mac/thread.cpp function
wxThread::Sleep when CLOCKS_PER_SEC is not
1000. The amount-to-sleep code doesn't take into
account this value. Due to this when I did Sleep(1000) it
was sleeping a fraction of a second because on my Mac
CLOCKS_PER_SEC is 60. This patch fixes it.

Dimitri Schoolwerth (dimitrishortcut)