Applied patch [ 608370 ] fix for static text sizing behaviour
wxStaticText is meant to resize when SetLabel is called
unless the flag wxST_NO_AUTORESIZE is used when
creating it. Motif currently ignores this flag and always
resizes the control.
The attached patch provides a wxStaticText specific
override of SetLabel() which honours the resize policy
requested via the wxST_NO_AUTORESIZE flag.

I have tested on AIX 4.3 which is Motif 2.1