Replace wxNSTimerData -setTimer method with -initWithWxTimer: initializer which correctly calls the superclass (NSObject) init.
Add an -init method for correctness; technically not needed as it is never
called but it's an Objective-C best practice.

Release the wxNSTimerData instance just after passing it to the NSTimer
factory method.  NSTimer retains it and we don't keep a pointer to it
so we should not keep a refence to it.  This fixes the bug in wxTimer::Stop
where the program crashes in the NSCFTimer userInfo method because the
NSTimer has already been invalidated.