Merged trunk 47580, 47581:
Rewrite conversion from classic Mac OS 'CURS'-style structures to NSCursor.
* Leave data as separate image and mask planes instead of interleaving it
* Specify bytesPerRow to make compatible with 10.4 (Tiger) and higher
* Use endian-independent math.
* Document that the cursor data came from wxMac source.

Rework the wxCusor ctor taking wx stock number to provide as many cursors as
possible by using the system cursors first, a set of built-in cursors copied 
from wxMac if a system cursor was not found, and a normal cursor otherwise.

Fixes lack of wxCURSOR_HAND and makes the code easier to follow by clearly
showing which cursors are and are not available from each source.

Copyright 2007, Software 2000 Ltd.