Fix for windres flags used in Win64 mingw build.
We must explicitly define WX_CPU_AMD64 when compiling using mingw64 and
configure as wx/msw/genrcdefs.h is not used in this case and even if it were,
it wouldn't work because g++ doesn't define _M_AMD64 (unlike MSVC).

So just add the flag directly to windres command line in configure.

Closes #11336.