Don't stop dragging in the grid when the mouse leaves the window.
Unexpected/unaccounted for mouse leaving and entering events stopped the drag
operation currently in progress in wxGrid. And while it was resumed later,
this resulted in the mouse being captured only twice but released only once.

Fix this by ignoring the leaving and entering events and checking that we
don't capture the mouse again.

Closes #11662.