Rework wxUIActionSimulator X11 backend.
Refactor key event sending into a separate function and move it into an anonymous namespace along with the existing SendButtonEvent. Replace much of the window and display code with a simpler alternative. Add conversion from a wxKeyCode to an X11 KeyCode as per wxMSW, to do this wxCharCodeXToWX and wxCharCodeWXToX declarations were moved to utilsx11.h and WXKeySym was defined.