Don't forbid creating wxSplitterWindow with border style.
Any border specified for wxSplitterWindow was explicitly discarded when
creating it but there doesn't seem to be any reason to forbid it, the original
code probably predated the addition of wxWindow::GetDefaultBorder() which
allowed to have different borders by default for different classes.

In any case, simply remove this code now to allow creating splitters with
borders if so desired.

Closes #12413.