Reset the checked state of the first radio item in wxAuiToolBar too.
Fix the condition of the loop in wxAuiToolBar::ToggleTool() to also uncheck
the item with index 0 that was always excluded before because of "i > 0" check
in the loop.

Also exclude the item being turned on itself from the loops, it's useless to
reset its checked state only to turn it on again below.

Closes #11110.