Update status bar labels properly after width change in wxMSW.
The labels were not updated (i.e. ellipsized or, on the contrary, displayed in
full) correctly after a status bar field was changed; there was even a FIXME
comment about this in the code.

Fix this by calling DoUpdateStatusText() when the field width is changed and
also update the field width after resetting the existing tooltips.