Fix choice of tree item icon for the selected state in wxMSW.
The logic for selecting the item icon was broken when selected image was
specified as it was not taken into account as fallback for selected expanded
state so selected icon was used only for collapsed items.

Fix this by using the selected icon if it is specified and selected expanded
one is not.